manual/requirements.txt
# Core libraries for screen automation
pyautogui==0.9.54

# Image processing - updated versions for Python 3.12 compatibility
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# OCR (Text recognition) - updated version
pytesseract==0.3.13

# Additional dependencies for Python 3.12
setuptools==69.0.3

# Web framework dependencies (if needed)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Platform specific
pywin32==306; sys_platform == "win32"
pyobjc-framework-LaunchServices==10.1; sys_platform == "darwin"

# Additional utilities
aiofiles==23.2.1
httpx==0.26.0
pydantic==2.5.3